Split global settings and app data into separate storage areas

+++ software/o_c_REV/POLYLFO.ino
@@ -76,7 +76,7 @@ void POLYLFO_init() {
   poly_lfo.Init();
 }
 
-static const size_t POLYLFO_SETTINGS_SIZE = sizeof(uint16_t) * POLYLFO_SETTING_LAST;
+static const size_t POLYLFO_SETTINGS_SIZE = sizeof(uint8_t) * POLYLFO_SETTING_LAST;
 
 size_t POLYLFO_save(char *storage) {
   return poly_lfo.save<uint8_t>(storage);
